+++ sandbox/vue-library/.storybook/main.ts
@@ -2,6 +2,7 @@ import type { StorybookConfig } from '@storybook/vue3-vite';
 
 const config: StorybookConfig = {
   stories: ['../src/stories/**/*.stories.@(js|jsx|mjs|ts|tsx|mdx)'],
+  staticDirs: [{ from: '../../../docs/assets', to: '/assets' }],
   addons: ['@storybook/addon-docs'],
   framework: {
     name: '@storybook/vue3-vite',