+++ sandbox/vue-library/.storybook/preview.ts
@@ -5,8 +5,12 @@ import '../../../packages/ui/dist/css/cedar-base.css';
 import '../../../packages/ui/dist/css/components/button.css';
 import '../../../packages/ui/dist/css/components/container.css';
 import '../../../packages/ui/dist/css/components/link.css';
+import { themes } from 'storybook/theming';
 
 export const parameters = {
+  docs: {
+    theme: themes.dark,
+  },
   controls: {
     expanded: true,
     matchers: {
@@ -15,3 +19,5 @@ export const parameters = {
     },
   },
 };
+
+export const tags = ['autodocs'];